Sped up the affine subpixel correlator by a factor of 2 by using pixel accessors to grab image pixels rather than the indexing operator.

+++ src/vw/Stereo/SubpixelView.h
@@ -63,7 +63,7 @@ namespace stereo {
     
     inline pixel_accessor origin() const { return pixel_accessor( *this, 0, 0 ); }
     
-    inline pixel_type operator()(double x, double y, int32 p = 0) const {
+    inline pixel_type operator()(float x, float y, int32 p = 0) const {
       vw_throw(NoImplErr() << "SubpixelView::operator() is not yet implemented.");
       return PixelDisparity<float>(); // Never reached
     }
